organizing the program
Aqui nós criamos a pasta "comunidadeimpressionadora" e colocamos todos os arquivos separadamente, com o arquivo __init__.py como base. Assim o programa fica organizado e evita problemas futuros de importações entre arquivos.
